[PodVariant] Ignore nil swift_version's in equality
This commit breaks the equality of PodVariant in order to make targets
which do not use Swift, able to share pods with ones that do.

This fixes an issue where there would be duplicate PodTargets created,
which would cause a collision on the filesystem in cases where there is
no swift version.